Voice AI Assistant by Developer Agrim Rana!
This project is a Python-based AI Voice Assistant that helps users perform a variety of everyday tasks using simple voice commands. The assistant is designed to be lightweight, user-friendly, privacy-focused, and highly functional. It uses text-to-speech (TTS) and speech recognition technologies to interact with the user through voice, making it a hands-free personal assistant.